remove Html module and make Html2 the default
