+++ lib/js/src/tea.js
@@ -10,8 +10,6 @@ var App;
 
 var Debug;
 
-var Html;
-
 var Html2;
 
 var Svg;
@@ -40,7 +38,6 @@ exports.Cmd = Cmd;
 exports.Sub = Sub;
 exports.App = App;
 exports.Debug = Debug;
-exports.Html = Html;
 exports.Html2 = Html2;
 exports.Svg = Svg;
 exports.Task = Task;
