--- lib/js/src/tea_html.js
@@ -1,1150 +1,0 @@
-// Generated by ReScript, PLEASE EDIT WITH CARE
-'use strict';
-
-var List = require("rescript/lib/js/list.js");
-var Vdom = require("./vdom.js");
-var Curry = require("rescript/lib/js/curry.js");
-var $$String = require("rescript/lib/js/string.js");
-var Tea_app = require("./tea_app.js");
-var Tea_json = require("./tea_json.js");
-var Caml_option = require("rescript/lib/js/caml_option.js");
-
-function text(str) {
-  return {
-          TAG: /* Text */1,
-          _0: str
-        };
-}
-
-var lazy1 = Vdom.lazyGen;
-
-function node(namespaceOpt, tagName, keyOpt, uniqueOpt, props, nodes) {
-  var namespace = namespaceOpt !== undefined ? namespaceOpt : "";
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode(namespace, tagName, key, unique, props, nodes);
-}
-
-function stopPropagation(t) {
-  t.stopPropagation();
-  
-}
-
-function preventDefault(t) {
-  t.preventDefault();
-  
-}
-
-function br(props) {
-  return Vdom.fullnode("", "br", "br", "br", props, /* [] */0);
-}
-
-function br$p(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "br", key, unique, props, nodes);
-}
-
-function div(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "div", key, unique, props, nodes);
-}
-
-function span(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "span", key, unique, props, nodes);
-}
-
-function p(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "p", key, unique, props, nodes);
-}
-
-function pre(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "pre", key, unique, props, nodes);
-}
-
-function a(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "a", key, unique, props, nodes);
-}
-
-function section(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "section", key, unique, props, nodes);
-}
-
-function header(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "header", key, unique, props, nodes);
-}
-
-function footer(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "footer", key, unique, props, nodes);
-}
-
-function h1(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h1", key, unique, props, nodes);
-}
-
-function h2(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h2", key, unique, props, nodes);
-}
-
-function h3(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h3", key, unique, props, nodes);
-}
-
-function h4(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h4", key, unique, props, nodes);
-}
-
-function h5(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h5", key, unique, props, nodes);
-}
-
-function h6(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "h6", key, unique, props, nodes);
-}
-
-function i(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "i", key, unique, props, nodes);
-}
-
-function strong(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "strong", key, unique, props, nodes);
-}
-
-function button(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "button", key, unique, props, nodes);
-}
-
-function input$p(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "input", key, unique, props, nodes);
-}
-
-function textarea(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "textarea", key, unique, props, nodes);
-}
-
-function label(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "label", key, unique, props, nodes);
-}
-
-function ul(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "ul", key, unique, props, nodes);
-}
-
-function ol(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "ol", key, unique, props, nodes);
-}
-
-function li(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "li", key, unique, props, nodes);
-}
-
-function table(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "table", key, unique, props, nodes);
-}
-
-function thead(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "thead", key, unique, props, nodes);
-}
-
-function tfoot(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "tfoot", key, unique, props, nodes);
-}
-
-function tbody(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "tbody", key, unique, props, nodes);
-}
-
-function th(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "th", key, unique, props, nodes);
-}
-
-function tr(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "tr", key, unique, props, nodes);
-}
-
-function td(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "td", key, unique, props, nodes);
-}
-
-function progress(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "progress", key, unique, props, nodes);
-}
-
-function img(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "img", key, unique, props, nodes);
-}
-
-function select(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "select", key, unique, props, nodes);
-}
-
-function option(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "option", key, unique, props, nodes);
-}
-
-function form(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "form", key, unique, props, nodes);
-}
-
-function nav(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "nav", key, unique, props, nodes);
-}
-
-function main(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "main", key, unique, props, nodes);
-}
-
-function aside(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "aside", key, unique, props, nodes);
-}
-
-function article(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "article", key, unique, props, nodes);
-}
-
-function details(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "details", key, unique, props, nodes);
-}
-
-function figcaption(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "figcaption", key, unique, props, nodes);
-}
-
-function figure(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "figure", key, unique, props, nodes);
-}
-
-function mark(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "mark", key, unique, props, nodes);
-}
-
-function summary(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "summary", key, unique, props, nodes);
-}
-
-function time(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "time", key, unique, props, nodes);
-}
-
-function hr(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "hr", key, unique, props, nodes);
-}
-
-function blockquote(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "blockquote", key, unique, props, nodes);
-}
-
-function code(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "code", key, unique, props, nodes);
-}
-
-function em(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "em", key, unique, props, nodes);
-}
-
-function b(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "b", key, unique, props, nodes);
-}
-
-function u(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "u", key, unique, props, nodes);
-}
-
-function sub(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "sub", key, unique, props, nodes);
-}
-
-function sup(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "sup", key, unique, props, nodes);
-}
-
-function dl(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "dl", key, unique, props, nodes);
-}
-
-function dt(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "dt", key, unique, props, nodes);
-}
-
-function dd(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "dd", key, unique, props, nodes);
-}
-
-function iframe(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "iframe", key, unique, props, nodes);
-}
-
-function canvas(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "canvas", key, unique, props, nodes);
-}
-
-function address(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "address", key, unique, props, nodes);
-}
-
-function caption(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "caption", key, unique, props, nodes);
-}
-
-function colgroup(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "colgroup", key, unique, props, nodes);
-}
-
-function col(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "col", key, unique, props, nodes);
-}
-
-function fieldset(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "fieldset", key, unique, props, nodes);
-}
-
-function legend(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "legend", key, unique, props, nodes);
-}
-
-function datalist(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "datalist", key, unique, props, nodes);
-}
-
-function optgroup(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "optgroup", key, unique, props, nodes);
-}
-
-function output(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "output", key, unique, props, nodes);
-}
-
-function meter(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "meter", key, unique, props, nodes);
-}
-
-function audio(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "audio", key, unique, props, nodes);
-}
-
-function video(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "video", key, unique, props, nodes);
-}
-
-function source(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "source", key, unique, props, nodes);
-}
-
-function track(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "track", key, unique, props, nodes);
-}
-
-function embed(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "embed", key, unique, props, nodes);
-}
-
-function object(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "object", key, unique, props, nodes);
-}
-
-function param(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "param", key, unique, props, nodes);
-}
-
-function ins(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "ins", key, unique, props, nodes);
-}
-
-function del(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "del", key, unique, props, nodes);
-}
-
-function small(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "small", key, unique, props, nodes);
-}
-
-function cite(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "cite", key, unique, props, nodes);
-}
-
-function dfn(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "dfn", key, unique, props, nodes);
-}
-
-function abbr(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "abbr", key, unique, props, nodes);
-}
-
-function $$var(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "var", key, unique, props, nodes);
-}
-
-function samp(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "samp", key, unique, props, nodes);
-}
-
-function kbd(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "kbd", key, unique, props, nodes);
-}
-
-function s(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "s", key, unique, props, nodes);
-}
-
-function q(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "q", key, unique, props, nodes);
-}
-
-function rt(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "rt", key, unique, props, nodes);
-}
-
-function bdi(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "bdi", key, unique, props, nodes);
-}
-
-function bdo(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "bdo", key, unique, props, nodes);
-}
-
-function wbr(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "wbr", key, unique, props, nodes);
-}
-
-function menuitem(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "menuitem", key, unique, props, nodes);
-}
-
-function menu(keyOpt, uniqueOpt, props, nodes) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  var unique = uniqueOpt !== undefined ? uniqueOpt : "";
-  return Vdom.fullnode("", "menu", key, unique, props, nodes);
-}
-
-function id(str) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "id",
-          _1: str
-        };
-}
-
-function href(str) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "href",
-          _2: str
-        };
-}
-
-function src(str) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "src",
-          _2: str
-        };
-}
-
-function title(str) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "title",
-          _2: str
-        };
-}
-
-function $$class(name) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "className",
-          _1: name
-        };
-}
-
-function classList(classes) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "className",
-          _1: $$String.concat(" ", List.map((function (param) {
-                      return param[0];
-                    }), List.filter(function (param) {
-                          return param[1];
-                        })(classes)))
-        };
-}
-
-function type$p(typ) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "type",
-          _1: typ
-        };
-}
-
-var style = Vdom.style;
-
-function styles(s) {
-  return {
-          TAG: /* Style */4,
-          _0: s
-        };
-}
-
-function placeholder(str) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "placeholder",
-          _1: str
-        };
-}
-
-function autofocus(b) {
-  if (b) {
-    return {
-            TAG: /* RawProp */0,
-            _0: "autofocus",
-            _1: "autofocus"
-          };
-  } else {
-    return /* NoProp */0;
-  }
-}
-
-function value(str) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "value",
-          _1: str
-        };
-}
-
-function name(str) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "name",
-          _1: str
-        };
-}
-
-function checked(b) {
-  if (b) {
-    return {
-            TAG: /* RawProp */0,
-            _0: "checked",
-            _1: "checked"
-          };
-  } else {
-    return /* NoProp */0;
-  }
-}
-
-function for$p(str) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "htmlFor",
-          _1: str
-        };
-}
-
-function hidden(b) {
-  if (b) {
-    return {
-            TAG: /* RawProp */0,
-            _0: "hidden",
-            _1: "hidden"
-          };
-  } else {
-    return /* NoProp */0;
-  }
-}
-
-function target(t) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "target",
-          _1: t
-        };
-}
-
-function action(a) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "action",
-          _1: a
-        };
-}
-
-function method(m) {
-  return {
-          TAG: /* RawProp */0,
-          _0: "method",
-          _1: m
-        };
-}
-
-var onCB = Vdom.onCB;
-
-var onMsg = Vdom.onMsg;
-
-function onInputOpt(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return Vdom.onCB("input", key, (function (ev) {
-                var target = ev.target;
-                if (target === undefined) {
-                  return ;
-                }
-                var value = target.value;
-                if (value !== undefined) {
-                  return Curry._1(msg, value);
-                }
-                
-              }));
-}
-
-function onInput(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return onInputOpt(key, (function (ev) {
-                return Caml_option.some(Curry._1(msg, ev));
-              }));
-}
-
-function onChangeOpt(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return Vdom.onCB("change", key, (function (ev) {
-                var target = ev.target;
-                if (target === undefined) {
-                  return ;
-                }
-                var value = target.value;
-                if (value !== undefined) {
-                  return Curry._1(msg, value);
-                }
-                
-              }));
-}
-
-function onChange(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return onChangeOpt(key, (function (ev) {
-                return Caml_option.some(Curry._1(msg, ev));
-              }));
-}
-
-function onClick(msg) {
-  return Vdom.onMsg("click", msg);
-}
-
-function onDoubleClick(msg) {
-  return Vdom.onMsg("dblclick", msg);
-}
-
-function onBlur(msg) {
-  return Vdom.onMsg("blur", msg);
-}
-
-function onFocus(msg) {
-  return Vdom.onMsg("focus", msg);
-}
-
-function onCheckOpt(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return Vdom.onCB("change", key, (function (ev) {
-                var target = ev.target;
-                if (target === undefined) {
-                  return ;
-                }
-                var value = target.checked;
-                if (value !== undefined) {
-                  return Curry._1(msg, value);
-                }
-                
-              }));
-}
-
-function onCheck(keyOpt, msg) {
-  var key = keyOpt !== undefined ? keyOpt : "";
-  return onCheckOpt(key, (function (ev) {
-                return Caml_option.some(Curry._1(msg, ev));
-              }));
-}
-
-function onMouseDown(msg) {
-  return Vdom.onMsg("mousedown", msg);
-}
-
-function onMouseUp(msg) {
-  return Vdom.onMsg("mouseup", msg);
-}
-
-function onMouseEnter(msg) {
-  return Vdom.onMsg("mouseenter", msg);
-}
-
-function onMouseLeave(msg) {
-  return Vdom.onMsg("mouseleave", msg);
-}
-
-function onMouseOver(msg) {
-  return Vdom.onMsg("mouseover", msg);
-}
-
-function onMouseOut(msg) {
-  return Vdom.onMsg("mouseout", msg);
-}
-
-var defaultOptions = {
-  stopPropagation: false,
-  preventDefault: false
-};
-
-function onWithOptions(key, eventName, options, decoder) {
-  return Vdom.onCB(eventName, key, (function ($$event) {
-                if (options.stopPropagation) {
-                  $$event.stopPropagation();
-                }
-                if (options.preventDefault) {
-                  $$event.preventDefault();
-                }
-                var result = Tea_json.Decoder.decodeEvent(decoder, $$event);
-                if (result.TAG === /* Ok */0) {
-                  return Caml_option.some(result._0);
-                }
-                
-              }));
-}
-
-function on(key, eventName, decoder) {
-  return onWithOptions(key, eventName, defaultOptions, decoder);
-}
-
-var targetValue = Tea_json.Decoder.at({
-      hd: "target",
-      tl: {
-        hd: "value",
-        tl: /* [] */0
-      }
-    }, Tea_json.Decoder.string);
-
-var targetChecked = Tea_json.Decoder.at({
-      hd: "target",
-      tl: {
-        hd: "checked",
-        tl: /* [] */0
-      }
-    }, Tea_json.Decoder.bool);
-
-var keyCode = Tea_json.Decoder.field("keyCode", Tea_json.Decoder.$$int);
-
-function max(value) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "max",
-          _2: value
-        };
-}
-
-function min(value) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "min",
-          _2: value
-        };
-}
-
-function step(value) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "step",
-          _2: value
-        };
-}
-
-function disabled(b) {
-  if (b) {
-    return {
-            TAG: /* Attribute */1,
-            _0: "",
-            _1: "disabled",
-            _2: "true"
-          };
-  } else {
-    return /* NoProp */0;
-  }
-}
-
-function selected(b) {
-  if (b) {
-    return {
-            TAG: /* Attribute */1,
-            _0: "",
-            _1: "selected",
-            _2: "true"
-          };
-  } else {
-    return /* NoProp */0;
-  }
-}
-
-function acceptCharset(c) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "accept-charset",
-          _2: c
-        };
-}
-
-function rel(value) {
-  return {
-          TAG: /* Attribute */1,
-          _0: "",
-          _1: "rel",
-          _2: value
-        };
-}
-
-var Attributes = {
-  max: max,
-  min: min,
-  step: step,
-  disabled: disabled,
-  selected: selected,
-  acceptCharset: acceptCharset,
-  rel: rel
-};
-
-var map = Tea_app.map;
-
-var noNode = Vdom.noNode;
-
-var noProp = /* NoProp */0;
-
-exports.map = map;
-exports.noNode = noNode;
-exports.text = text;
-exports.lazy1 = lazy1;
-exports.node = node;
-exports.stopPropagation = stopPropagation;
-exports.preventDefault = preventDefault;
-exports.br = br;
-exports.br$p = br$p;
-exports.div = div;
-exports.span = span;
-exports.p = p;
-exports.pre = pre;
-exports.a = a;
-exports.section = section;
-exports.header = header;
-exports.footer = footer;
-exports.h1 = h1;
-exports.h2 = h2;
-exports.h3 = h3;
-exports.h4 = h4;
-exports.h5 = h5;
-exports.h6 = h6;
-exports.i = i;
-exports.strong = strong;
-exports.button = button;
-exports.input$p = input$p;
-exports.textarea = textarea;
-exports.label = label;
-exports.ul = ul;
-exports.ol = ol;
-exports.li = li;
-exports.table = table;
-exports.thead = thead;
-exports.tfoot = tfoot;
-exports.tbody = tbody;
-exports.th = th;
-exports.tr = tr;
-exports.td = td;
-exports.progress = progress;
-exports.img = img;
-exports.select = select;
-exports.option = option;
-exports.form = form;
-exports.nav = nav;
-exports.main = main;
-exports.aside = aside;
-exports.article = article;
-exports.details = details;
-exports.figcaption = figcaption;
-exports.figure = figure;
-exports.mark = mark;
-exports.summary = summary;
-exports.time = time;
-exports.hr = hr;
-exports.blockquote = blockquote;
-exports.code = code;
-exports.em = em;
-exports.b = b;
-exports.u = u;
-exports.sub = sub;
-exports.sup = sup;
-exports.dl = dl;
-exports.dt = dt;
-exports.dd = dd;
-exports.iframe = iframe;
-exports.canvas = canvas;
-exports.address = address;
-exports.caption = caption;
-exports.colgroup = colgroup;
-exports.col = col;
-exports.fieldset = fieldset;
-exports.legend = legend;
-exports.datalist = datalist;
-exports.optgroup = optgroup;
-exports.output = output;
-exports.meter = meter;
-exports.audio = audio;
-exports.video = video;
-exports.source = source;
-exports.track = track;
-exports.embed = embed;
-exports.object = object;
-exports.param = param;
-exports.ins = ins;
-exports.del = del;
-exports.small = small;
-exports.cite = cite;
-exports.dfn = dfn;
-exports.abbr = abbr;
-exports.$$var = $$var;
-exports.samp = samp;
-exports.kbd = kbd;
-exports.s = s;
-exports.q = q;
-exports.rt = rt;
-exports.bdi = bdi;
-exports.bdo = bdo;
-exports.wbr = wbr;
-exports.menuitem = menuitem;
-exports.menu = menu;
-exports.noProp = noProp;
-exports.id = id;
-exports.href = href;
-exports.src = src;
-exports.title = title;
-exports.$$class = $$class;
-exports.classList = classList;
-exports.type$p = type$p;
-exports.style = style;
-exports.styles = styles;
-exports.placeholder = placeholder;
-exports.autofocus = autofocus;
-exports.value = value;
-exports.name = name;
-exports.checked = checked;
-exports.for$p = for$p;
-exports.hidden = hidden;
-exports.target = target;
-exports.action = action;
-exports.method = method;
-exports.onCB = onCB;
-exports.onMsg = onMsg;
-exports.onInputOpt = onInputOpt;
-exports.onInput = onInput;
-exports.onChangeOpt = onChangeOpt;
-exports.onChange = onChange;
-exports.onClick = onClick;
-exports.onDoubleClick = onDoubleClick;
-exports.onBlur = onBlur;
-exports.onFocus = onFocus;
-exports.onCheckOpt = onCheckOpt;
-exports.onCheck = onCheck;
-exports.onMouseDown = onMouseDown;
-exports.onMouseUp = onMouseUp;
-exports.onMouseEnter = onMouseEnter;
-exports.onMouseLeave = onMouseLeave;
-exports.onMouseOver = onMouseOver;
-exports.onMouseOut = onMouseOut;
-exports.defaultOptions = defaultOptions;
-exports.onWithOptions = onWithOptions;
-exports.on = on;
-exports.targetValue = targetValue;
-exports.targetChecked = targetChecked;
-exports.keyCode = keyCode;
-exports.Attributes = Attributes;
-/* targetValue Not a pure module */
