+++ lib/js/src/tea_html2.js
@@ -6,7 +6,6 @@ var Vdom = require("./vdom.js");
 var Curry = require("rescript/lib/js/curry.js");
 var $$String = require("rescript/lib/js/string.js");
 var Tea_app = require("./tea_app.js");
-var Tea_html = require("./tea_html.js");
 var Tea_json = require("./tea_json.js");
 var Caml_option = require("rescript/lib/js/caml_option.js");
 var Caml_string = require("rescript/lib/js/caml_string.js");
@@ -1654,10 +1653,53 @@ var onCB = Vdom.onCB;
 
 var onMsg = Vdom.onMsg;
 
+var defaultOptions = {
+  stopPropagation: false,
+  preventDefault: false
+};
+
+function onWithOptions(key, eventName, options, decoder) {
+  return Vdom.onCB(eventName, key, (function ($$event) {
+                if (options.stopPropagation) {
+                  $$event.stopPropagation();
+                }
+                if (options.preventDefault) {
+                  $$event.preventDefault();
+                }
+                var result = Tea_json.Decoder.decodeEvent(decoder, $$event);
+                if (result.TAG === /* Ok */0) {
+                  return Caml_option.some(result._0);
+                }
+                
+              }));
+}
+
+function on(key, eventName, decoder) {
+  return onWithOptions(key, eventName, defaultOptions, decoder);
+}
+
+var targetValue = Tea_json.Decoder.at({
+      hd: "target",
+      tl: {
+        hd: "value",
+        tl: /* [] */0
+      }
+    }, Tea_json.Decoder.string);
+
+var targetChecked = Tea_json.Decoder.at({
+      hd: "target",
+      tl: {
+        hd: "checked",
+        tl: /* [] */0
+      }
+    }, Tea_json.Decoder.bool);
+
+var keyCode = Tea_json.Decoder.field("keyCode", Tea_json.Decoder.$$int);
+
 function preventDefaultOn(keyOpt, eventName, decoder) {
   var key = keyOpt !== undefined ? keyOpt : "";
-  return Tea_html.onWithOptions(key, eventName, {
-              stopPropagation: Tea_html.defaultOptions.stopPropagation,
+  return onWithOptions(key, eventName, {
+              stopPropagation: false,
               preventDefault: true
             }, decoder);
 }
@@ -1772,15 +1814,19 @@ function onFocus(msg) {
   return Vdom.onMsg("focus", msg);
 }
 
+var map = Tea_app.map;
+
+var noNode = Vdom.noNode;
+
 var Events = {
   onCB: onCB,
   onMsg: onMsg,
-  on: Tea_html.on,
-  onWithOptions: Tea_html.onWithOptions,
-  defaultOptions: Tea_html.defaultOptions,
-  targetValue: Tea_html.targetValue,
-  targetChecked: Tea_html.targetChecked,
-  keyCode: Tea_html.keyCode,
+  on: on,
+  onWithOptions: onWithOptions,
+  defaultOptions: defaultOptions,
+  targetValue: targetValue,
+  targetChecked: targetChecked,
+  keyCode: keyCode,
   preventDefaultOn: preventDefaultOn,
   onClick: onClick,
   onDoubleClick: onDoubleClick,
@@ -1800,10 +1846,6 @@ var Events = {
   onBlur: onBlur,
   onFocus: onFocus
 };
-
-var map = Tea_app.map;
-
-var noNode = Vdom.noNode;
 
 exports.map = map;
 exports.text = text;
@@ -1915,4 +1957,4 @@ exports.title = title;
 exports.link = link;
 exports.Attributes = Attributes;
 exports.Events = Events;
-/* Tea_html Not a pure module */
+/* targetValue Not a pure module */
