+++ test/test_client_btn_update_span.res
@@ -1,5 +1,7 @@
 open Tea.App
-open Tea.Html
+open Tea_html2.Events 
+open Tea_html2.Attributes
+open Tea.Html2
 
 @deriving({accessors: accessors}) type msg = Trigger
