+++ test/test_client_counter_debug_beginner.res
@@ -1,5 +1,6 @@
 open Tea.App
-open Tea.Html
+open Tea.Html2
+open Tea_html2.Events 
 
 type msg =
   | Increment
@@ -29,7 +30,7 @@ let view = model =>
   div(
     list{},
     list{
-      span(list{style("text-weight", "bold")}, list{text(string_of_int(model))}),
+      span(list{Tea_html2.Attributes.style("text-weight", "bold")}, list{text(string_of_int(model))}),
       br(list{}),
       view_button(
         "Increment",
