+++ test/test_client_http_task.res
@@ -1,5 +1,6 @@
 open Tea
-open Tea.Html
+open Tea.Html2
+open Tea_html2.Events 
 type result<'ok, 'err> = result<'ok, 'err>
 
 @deriving(accessors)
